+++ tests/tools/emr/__init__.py
@@ -42,7 +42,7 @@ class ToolTestCase(MockBotoTestCase):
         self.add_mock_s3_data({'walrus': {}})
         kwargs.update(dict(
             conf_paths=[],
-            s3_tmp_dir='s3://walrus/',
+            cloud_tmp_dir='s3://walrus/',
             s3_sync_wait_time=0))
         with EMRJobRunner(**kwargs) as runner:
             return runner.make_persistent_cluster()