+++ tests/tools/emr/test_terminate_idle_clusters.py
@@ -362,7 +362,7 @@ class ClusterTerminationTestCase(MockBotoTestCase):
         if 'now' not in kwargs:
             kwargs['now'] = self.now
 
-        kwargs['s3_tmp_dir'] = 's3://my_bucket/'
+        kwargs['cloud_tmp_dir'] = 's3://my_bucket/'
         kwargs['s3_sync_wait_time'] = 0
         kwargs['max_mins_locked'] = 1